cmake: Moved finding and running CMake out of dependency
